backend/requirements.txt
Flask~=3.0.3
Flask-CORS~=6.0.0
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
openai-whisper==20250625
gunicorn==23.0.0
python-dotenv==1.1.1
Werkzeug~=3.1.3
pytest~=7.4.4
